GDGT-2194 move useIcon (and the related hooks) to the hooks folder
