+++ enterprise/frontend/src/metabase-enterprise/data-studio/library/LibraryPage/hooks.ts
@@ -9,8 +9,8 @@ import type {
   TreeItem,
 } from "metabase/data-studio/common/types";
 import { createEmptyStateItem } from "metabase/data-studio/common/utils";
+import { useGetIcon } from "metabase/hooks/use-icon";
 import { useMetadataToasts } from "metabase/metadata/hooks";
-import { useGetIcon } from "metabase/utils/icon";
 import { useSelector } from "metabase/utils/redux";
 import { getIsRemoteSyncReadOnly } from "metabase-enterprise/remote_sync/selectors";
 import type { Collection, CollectionId } from "metabase-types/api";
