+++ enterprise/frontend/src/metabase-enterprise/remote_sync/components/ChangesLists/EntityLink.tsx
@@ -1,7 +1,8 @@
 import { useMemo } from "react";
 
+import { useGetIcon } from "metabase/hooks/use-icon";
 import { Anchor, Group, Icon } from "metabase/ui";
-import { type IconData, type IconModel, useGetIcon } from "metabase/utils/icon";
+import type { IconData, IconModel } from "metabase/utils/icon";
 import { modelToUrl } from "metabase/utils/urls";
 import type { RemoteSyncEntity } from "metabase-types/api";
