+++ frontend/src/metabase/dashboard/components/QuestionPicker/QuestionPicker.tsx
@@ -13,6 +13,7 @@ import { useDashboardContext } from "metabase/dashboard/context";
 import { getDashboard } from "metabase/dashboard/selectors";
 import { isEmbeddingSdk } from "metabase/embedding-sdk/config";
 import { Collections, ROOT_COLLECTION } from "metabase/entities/collections";
+import { useGetIcon } from "metabase/hooks/use-icon";
 import { PLUGIN_COLLECTIONS } from "metabase/plugins";
 import {
   canUserCreateNativeQueries,
@@ -20,7 +21,6 @@ import {
 } from "metabase/selectors/user";
 import { Button, Flex, Icon } from "metabase/ui";
 import { SEARCH_DEBOUNCE_DURATION } from "metabase/utils/constants";
-import { useGetIcon } from "metabase/utils/icon";
 import { connect, useDispatch, useSelector } from "metabase/utils/redux";
 import type { Collection, CollectionId } from "metabase-types/api";
