+++ frontend/src/metabase/home/components/HomePopularSection/HomePopularSection.tsx
@@ -3,7 +3,7 @@ import _ from "underscore";
 
 import { useListPopularItemsQuery } from "metabase/api";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
-import { useGetIcon } from "metabase/utils/icon";
+import { useGetIcon } from "metabase/hooks/use-icon";
 import { getName } from "metabase/utils/name";
 import * as Urls from "metabase/utils/urls";
 import type { PopularItem } from "metabase-types/api";
