+++ frontend/src/metabase/home/components/HomeRecentSection/HomeRecentSection.tsx
@@ -2,8 +2,8 @@ import { t } from "ttag";
 
 import { useListRecentsQuery } from "metabase/api";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
+import { useGetIcon } from "metabase/hooks/use-icon";
 import { getUser } from "metabase/selectors/user";
-import { useGetIcon } from "metabase/utils/icon";
 import { getName } from "metabase/utils/name";
 import { useSelector } from "metabase/utils/redux";
 import * as Urls from "metabase/utils/urls";
