+++ frontend/src/metabase/hooks/use-icon.ts
@@ -3,11 +3,10 @@ import { useCallback } from "react";
 import { getLibraryCollectionType } from "metabase/data-studio/utils";
 import { PERSONAL_COLLECTIONS } from "metabase/entities/collections/constants";
 import { PLUGIN_COLLECTIONS, PLUGIN_CUSTOM_VIZ } from "metabase/plugins";
+import type { IconData, ObjectWithModel } from "metabase/utils/icon";
+import { modelIconMap } from "metabase/utils/icon";
 import { getIconForVisualizationType } from "metabase/visualizations";
 import type { CollectionType, VisualizationDisplay } from "metabase-types/api";
-
-import type { IconData, ObjectWithModel } from "./icon";
-import { modelIconMap } from "./icon";
 
 export const useGetIconForVisualizationType = () => {
   const getCustomVizIcon = PLUGIN_CUSTOM_VIZ.useCustomVizPluginsIcon();
