+++ frontend/src/metabase/nav/containers/MainNavbar/MainNavbarContainer/BookmarkList/BookmarkList.tsx
@@ -15,9 +15,9 @@ import { CollapseSection } from "metabase/common/components/CollapseSection";
 import { Sortable } from "metabase/common/components/Sortable";
 import GrabberS from "metabase/css/components/grabber.module.css";
 import { Bookmarks } from "metabase/entities/bookmarks";
+import { useGetIcon } from "metabase/hooks/use-icon";
 import { PLUGIN_COLLECTIONS } from "metabase/plugins";
 import { Icon, Tooltip } from "metabase/ui";
-import { useGetIcon } from "metabase/utils/icon";
 import { connect } from "metabase/utils/redux";
 import * as Urls from "metabase/utils/urls";
 import type { Bookmark } from "metabase-types/api";
