+++ frontend/src/metabase/query_builder/components/view/sidebars/QuestionInfoSidebar/components/TablesLinkedToQuestion.tsx
@@ -2,9 +2,9 @@ import { useMemo } from "react";
 import { t } from "ttag";
 
 import { Link } from "metabase/common/components/Link";
+import { useGetIcon } from "metabase/hooks/use-icon";
 import { getQuestionWithoutComposing } from "metabase/query_builder/selectors";
 import { Flex, Icon, Stack, Text } from "metabase/ui";
-import { useGetIcon } from "metabase/utils/icon";
 import { useSelector } from "metabase/utils/redux";
 
 import { ToggleFullList } from "./ToggleFullList";
