+++ frontend/src/metabase/questions/components/CollectionBadge.tsx
@@ -2,9 +2,9 @@ import type { ComponentType, PropsWithChildren } from "react";
 
 import { Badge } from "metabase/common/components/Badge";
 import { Collections } from "metabase/entities/collections";
+import { useGetIcon } from "metabase/hooks/use-icon";
 import { useTranslateContent } from "metabase/i18n/hooks";
 import { PLUGIN_COLLECTIONS } from "metabase/plugins";
-import { useGetIcon } from "metabase/utils/icon";
 import { modelToUrl } from "metabase/utils/urls/modelToUrl";
 import type {
   CollectionId,
