+++ frontend/src/metabase/rich_text_editing/tiptap/extensions/SmartLink/SmartLinkNode.tsx
@@ -23,17 +23,14 @@ import {
 import { EntityIcon } from "metabase/common/components/EntityIcon";
 import { Link } from "metabase/common/components/Link";
 import { updateMentionsCache } from "metabase/documents/documents.slice";
+import { useGetIcon } from "metabase/hooks/use-icon";
 import {
   METABSE_PROTOCOL_MD_LINK,
   parseMetabaseProtocolMarkdownLink,
 } from "metabase/metabot/utils/links";
 import { PLUGIN_TRANSFORMS } from "metabase/plugins";
 import { Icon } from "metabase/ui";
-import {
-  type IconModel,
-  type ObjectWithModel,
-  useGetIcon,
-} from "metabase/utils/icon";
+import type { IconModel, ObjectWithModel } from "metabase/utils/icon";
 import { useDispatch } from "metabase/utils/redux";
 import { modelToUrl } from "metabase/utils/urls/modelToUrl";
 import { extractEntityId } from "metabase/utils/urls/utils";
